+++ transform/src/coverage.ts
@@ -39,6 +39,8 @@ const COVERAGE_IGNORED_CALLS = new Set([
   "unmockFn",
   "mockImport",
   "unmockImport",
+  "snapshotImport",
+  "restoreImport",
 ]);
 
 class CoverPoint {